issue-24: Finished entwine functional test

+++ usecase/entwine/scripts/run-dual-binge-entwine.sh
@@ -1,10 +1,12 @@
 ROOTDIR=$( dirname $0 )/../../..
 
 function cleanup {
-    docker kill dynamodb
-    docker rm dynamodb
-    docker kill minio
-    docker rm minio
+    if [[ -z ${CI_TEST} ]]; then
+        docker kill dynamodb
+        docker rm dynamodb
+        docker kill minio
+        docker rm minio
+    fi
 }
 
 trap cleanup EXIT INT TERM
@@ -13,12 +15,17 @@ export AWS_ACCESS_KEY_ID=localtest
 export AWS_SECRET_ACCESS_KEY=localtest
 export AWS_DEFAULT_REGION=us-west-2
 
+mkdir -p /tmp/authenticators
+mkdir -p /tmp/a-out
+mkdir -p /tmp/b-out
+
 rm /tmp/a-out/*
 rm /tmp/b-out/*
 
-docker run -d -p 8000:8000 --name dynamodb amazon/dynamodb-local
-
-sleep 2
+if [[ -z ${CI_TEST} ]]; then
+    docker run -d -p 8000:8000 --name dynamodb amazon/dynamodb-local -jar DynamoDBLocal.jar -sharedDb
+    sleep 2
+fi
 
 # Create tables 
 aws dynamodb create-table --endpoint-url http://localhost:8000 --region us-west-2 \
@@ -45,15 +52,17 @@ aws dynamodb create-table --endpoint-url http://localhost:8000 --region us-west-
      --key-schema AttributeName=KeyPrefix,KeyType=HASH AttributeName=ValueKey,KeyType=RANGE \
      --provisioned-throughput ReadCapacityUnits=1,WriteCapacityUnits=1
 
-docker run -d -p 9000:9000 \
-  -e "MINIO_ROOT_USER=localtest" \
-  -e "MINIO_ROOT_PASSWORD=localtest" \
-  --name minio \
-  minio/minio server /data
+if [[ -z ${CI_TEST} ]]; then
+    docker run -d -p 9000:9000 \
+        -e "MINIO_ROOT_USER=localtest" \
+        -e "MINIO_ROOT_PASSWORD=localtest" \
+        --name minio \
+        minio/minio server /data
 
-sleep 2
+    sleep 2
+    mc alias set localtest http://localhost:9000 localtest localtest
+fi
 
-mc alias set localtest http://localhost:9000 localtest localtest
 
 # Create buckets to use in the test
 mc mb localtest/localtesta
@@ -67,22 +76,31 @@ rm /tmp/ticker.pem*; ssh-keygen -f /tmp/ticker.pem -t rsa -m pem -N "" && ssh-ke
 
 aws dynamodb list-tables --endpoint-url http://localhost:8000 --region us-west-2
 
-${ROOTDIR}/bin/ticker  --grpcPort 8001 --httpPort 8002 -kvConnectionString "dynamo:endpoint=http://localhost:8000,region=us-west-2,tableName=tickerKVStore,prefixLength=4" \
-    -authenticatorPath /tmp/authenticators -privateKeyPath /tmp/ticker.pem > /tmp/ticker.out 2>&1 &
-TICKER_PID=$!
+if [[ -n ${DEBUG_TICKER} ]]; then
+    read -p "Start debugger for ticker, then press any key to continue: " NOTHING
+else
+    ${ROOTDIR}/bin/ticker  --grpcPort 8001 --httpPort 8002 -kvConnectionString "dynamo:endpoint=http://localhost:8000,region=us-west-2,tableName=tickerKVStore,prefixLength=4" \
+        -authenticatorPath /tmp/authenticators -privateKeyPath /tmp/ticker.pem > /tmp/ticker.out 2>&1 &
+    TICKER_PID=$!
+fi
 
 sleep 2
 
-curl -X POST http://localhost:8002/api/v1/tick
+curl --silent -X POST http://localhost:8002/api/v1/tick > /dev/null 2>&1
 
-${ROOTDIR}/bin/binge  -daemonPath /entwine -daemonPort 8008 -maxConnections 16 -runType stateless-daemon -exporter stdout -config ${ROOTDIR}/usecase/entwine/config/binge-a.json > /tmp/binge-a.out 2>&1 &
-BINGE1_PID=$!
+if [[ -n ${DEBUG_BINGE} ]]; then
+    read -p "Start debugger for binge, then press any key to continue: " NOTHING
+else
+    ${ROOTDIR}/bin/binge  -daemonPath /entwine -daemonPort 8008 -maxConnections 16 -runType stateless-daemon -exporter stdout -config ${ROOTDIR}/usecase/entwine/config/binge-a.json > /tmp/binge-a.out 2>&1 &
+    BINGE1_PID=$!
+fi
+
 ${ROOTDIR}/bin/binge  -daemonPath /entwine -daemonPort 8009 -maxConnections 16 -runType stateless-daemon -exporter stdout -config ${ROOTDIR}/usecase/entwine/config/binge-b.json > /tmp/binge-b.out 2>&1 &
 BINGE2_PID=$!
 sleep 2
 
 
-let NUM_MESSAGES=30
+let NUM_MESSAGES=100
 
 let i=0
 let A_index=1
@@ -90,12 +108,13 @@ let B_index=1
 declare -a INDEXES
 while (( ${i} < ${NUM_MESSAGES} )); do
     choose=$(($RANDOM % 2))
-    payload='{"idx":'${i}'}'
     if (( ${choose} == 0 )); then
+        payload='{"idx":'${i}', "aIdx": '${A_index}'}'
         echo ${payload} | curl --silent -H "Content-Type: application/json" -X POST --data-binary @- http://localhost:8008/entwine -o /dev/null
         INDEXES[${i}]=${A_index}
         let A_index=${A_index}+1
     else
+        payload='{"idx":'${i}', "bIdx": '${B_index}'}'
         echo ${payload} | curl --silent -H "Content-Type: application/json" -X POST --data-binary @- http://localhost:8009/entwine -o /dev/null
         INDEXES[${i}]=${B_index}
         let B_index=${B_index}+1
@@ -110,10 +129,27 @@ for outfile in `echo /tmp/a-out/* /tmp/b-out/*`; do
     subStreamID=`cat ${outfile} | jq -r '.["internal:entwine:output"][0].subStreamID'`
     MESSAGES[${idx}]="${subStreamID}:${uuid}:"${INDEXES[${idx}]}
 done
-
-while (( ${i} > 20 )); do
-    echo ${ROOTDIR}/bin/entwinectl  -tickerEndpoint localhost:8001 -command HappenedBefore ${MESSAGES[$((i))]} ${MESSAGES[$((i-1))]}
+    
+let num_success=0
+let i=${i}-1
+while (( ${i} > 0 )); do
+    result=`${ROOTDIR}/bin/entwinectl  -tickerEndpoint localhost:8001 -command HappenedBefore ${MESSAGES[$((i))]} ${MESSAGES[$((i-1))]} | egrep -v -E '[0-9a-z]{8}\-[0-9a-z]{4}\-[0-9a-z]{4}\-[0-9a-z]{4}\-[0-9a-z]{12} happenedBefore [0-9a-z]{8}\-[0-9a-z]{4}\-[0-9a-z]{4}\-[0-9a-z]{4}\-[0-9a-z]{12}: false'`
+    if [[ -z ${result} ]]; then
+        let num_success=${num_success}+1
+    else
+        echo "Error evaluating happened before relationship: ${result}"
+    fi
     let i=${i}-1
 done
 
+ret=0
+if (( ${num_success} == ${NUM_MESSAGES} - 1 )); then
+    echo "RESULT: success"
+else
+    echo "RESULT: failure"
+    ret=1
+fi
+
 kill -9 ${TICKER_PID} ${BINGE1_PID} ${BINGE2_PID}
+
+exit ${ret}
